pyproject.toml
[project]
name = "pixels-pygame-palace"
version = "0.1.0"
description = "Flask backend for Pixel's PyGame Palace"
authors = [
    { name = "Jon Bogaty", email = "jon@jonbogaty.com" }
]
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "flask-limiter>=3.13",
    "flask-socketio>=5.5.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "psutil>=7.1.0",
    "pygame-ce>=2.5.5",
    "pygbag>=0.8.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-engineio>=4.12.2",
    "python-socketio>=5.13.0",
    "requests>=2.32.5",
    "rich>=14.1.0",
]

[project.scripts]
backend = "src.backend:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"

[tool.uv]
# UV configuration
